[BUGFIX] Use temporary file before adding to FAL
The upload script put the temporary upload file into a .part file at the
same location as the target file. This is now changed to using a
typo3temp/ based file.

Instead of the upload folder, the temp file path is now stored in the
session. This might pose problems when multiple files are uploaded in
parallel; if we ever need this, the code must be adjusted accordingly.
After the upload was successful, the file is added to FAL and the
temporary path removed from the session.